settings file format is changed to json to make it human readable
