+++ src/SQLQueryStress/SQLQueryStress.csproj
@@ -73,6 +73,10 @@
       <HintPath>.\ICSharpCode.AvalonEdit.dll</HintPath>
       <Private>False</Private>
     </Reference>
+    <Reference Include="Newtonsoft.Json, Version=6.0.0.0, Culture=neutral, PublicKeyToken=30ad4fe6b2a6aeed, processorArchitecture=MSIL">
+      <SpecificVersion>False</SpecificVersion>
+      <HintPath>.\Newtonsoft.Json.dll</HintPath>
+    </Reference>
     <Reference Include="PresentationCore" />
     <Reference Include="PresentationFramework" />
     <Reference Include="System" />
@@ -121,6 +125,7 @@
     <Compile Include="CommandLineParser\Text\HelpText.cs" />
     <Compile Include="CommandLineParser\Util\ReflectionUtil.cs" />
     <Compile Include="CommandLineParser\Util\Validator.cs" />
+    <Compile Include="ConnectionInfo.cs" />
     <Compile Include="DatabaseSelect.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -154,6 +159,7 @@
     </Compile>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="QueryStressSettings.cs" />
     <Compile Include="SqlControl.xaml.cs">
       <DependentUpon>SqlControl.xaml</DependentUpon>
     </Compile>
@@ -224,6 +230,7 @@
   </ItemGroup>
   <ItemGroup>
     <EmbeddedResource Include="ICSharpCode.AvalonEdit.dll" />
+    <EmbeddedResource Include="Newtonsoft.Json.dll" />
     <Content Include="Resources\QueryStressIcon.ico" />
   </ItemGroup>
   <ItemGroup>
